requirements.txt
click==8.1.7
openai==1.55.3
google-search-results==2.4.2
rich==13.9.4
python-dotenv==1.0.1
requests==2.32.3